ci(workflows): remove superfluous check.yaml
Looks like the current implementation of the workflows works well
enough.

Now the goal is to make them work a bit better, and be a bit more
maintainable. They currently share a lot of code, and that can either be
packaged into a module, or it can be assumed that each subproject file
will always be run from the main checks.py, and that, for instance, pipx
will always be installed.

In general, I'm not a fan of either of these, and I think it'd be better
to ensure they both have robust methods for doing the tasks they need,
and that if they both perform the same task, they do it the same way.

Currently, the pipx installation handling is less robust in
copy_unique/checks.py

The main problem with pipx was where it was installing the shims.
Currently, it's in a place that leads me to believe there's a bug in
pipx:

/home/runner/work/utilities/utilities/"/opt/pipx_bin"

With the double quotes being part of the path (i.e. there's a folder
name " or "/opt/pipx_bin").

Also, on Windows, the directory:
C:\Program Files (x86)\pipx_bin

is used, and I didn't know pip could just write to
C:\Program Files (x86)

Now that it's at least consistently discoverable, the files can be
cleaned up, and proper logging and CI detection can be used.

It works as-is, but because of the way it's doing printing, error
messages and tracebacks can become jumbled together with the rest of the
output.

Also, a major point will be optimizing this:

Caching will be a later step, as that can be hard to get right, and
harder to track down the platform-specific places where files are
stored.

Instead, narrowing down the files run and checked to only what's changed
will really help.

Also, currently, if one of the subproject checks fails, the rest aren't
checked. Ideally, each would contribute a check-mark, but for now it can
be collected into an all-or-nothing status.

Either way, error reporting needs to be improved to better indicate
where a failure occured.