feat: add minimum functioning project
I think implementing --hash can wait until GitHub pages have been set
up.

GitHub Actions should be able to automate using this tool to generate an
index that can then be committed to a `gh-pages` branch, so that no one
has to directly use this tool, and can just use the generated index
hosted by GitHub pages.

To keep the index up to date, I'll set the actions to run once a month.
In case someone needs it earlier than that, I'd like to be able to
trigger a run in response to someone interacting with the repository
through normal GitHub channels.

That way, if someone see's it's out of date, an issue can be opened, or
a stickied issue can be commented on, and the actions will trigger
regenerating the index.

To make sure it functions, I want to make that the only way I get the
index generated from the very first generation.