refactor(types): satisfy pylint and mypy
I don't like how this turned out. Ideally, Directory and OpenablePath
would be subclasses of pathlib.Path that implement stricter requirements
for the path, but subclassing Path is difficult:
https://bugs.python.org/issue24132

The implementation before this, that made the classes callable, and
returned Path objects on instantiation, worked for runtime, but made
mypy very unhappy.

I think the only way to cleanly implement this is to either subclass
str and implement the methods needed (like how StrictPath currently
looks), or, the direction I'm probably going to go: writing functions
that verify the desired behaviour, and calling them when necessary.